SVG.viewing(RSVGTipsDevice)	R Documentation

Ways to view SVG files

Description

The easiest way to view SVG files is in a web browser that has native SVG support. Mozilla Firefox (>= 1.5) and Opera (>= 9) both have this. Safari should have this soon (as of May 2007).

The section "Necessary Viewer/Browser for the examples" at the bottom of http://www.carto.net/papers/svg/samples/ appeared comprehensive and up-to-date as of May 2007.

Opera notes

The Opera viewer has automatic tooltips, which get data from the "title" element. When viewing SVG files created by RSVGTipsDevice, an extra tooltip will pop up after a short wait. To enable or disable tooltip popups in Opera, go to Tools > Preferences > Advanced > Browsing and toggle the "Show tooltips" option. (see http://www.opera.com/support/search/view/714/)

[Package RSVGTipsDevice version 0.7-5 Index]