getSVGToolTipMode(RSVGTipsDevice)	R Documentation

Return the tool tip mode of the RSVGTips device

Description

Return the tool tip mode of the RSVGTips device.

Usage


getSVGToolTipMode()

Details

Returns the mode that was set in the call to devSVGTips (this mode cannot be changed while a plot is being created.)

Value

An integer 	 representing the tooltip mode of the current device:

-1: the current device is not an RSVGTips device
0: the tooltips are not being used
1: tool tips are title + one line of description
2: tool tips are title + two lines of description

Author(s)

Tony Plate tplate@acm.org

See Also

devSVGTips

[Package RSVGTipsDevice version 0.8-1 Index]