setSVGShapeContents(RSVGTipsDevice)	R Documentation

Set raw shape contents (XML text) for the next graphics shape drawn with the RSVGTips device.

Description

Set raw shape contents (XML text) for the next graphics shape drawn with the RSVGTips device. This function provides lower-level access than setSVGShapeToolTip.

Usage


setSVGShapeContents(contents)

Arguments

contents 	 A character vector containing XML text.

Details

This function sets text that will be included in the SVG commands output for the next graphics shape drawn with RSVGTipsDevice.

Note that contents is used as-is, with no substitutions made for special characters. Consequently, it is possible to produce an invalid XML file by using special characters in contents (i.e., "<", ">", "&", and single- and double-quotes.) Such special characters should be replaced by their XML equivalent entity – see setSVGShapeToolTip for a list of these.

Value

Returns an invisible NULL.

Author(s)

Tony Plate tplate@acm.org

See Also

RSVGTipsDevice

[Package RSVGTipsDevice version 0.7-5 Index]