setSVGShapeToolTip(RSVGTipsDevice)	R Documentation

Set the tool tip for the next graphics shape drawn with the RSVGTips device.

Description

Set the tool tip for the next graphics shape drawn with the RSVGTips device.

Usage


setSVGShapeToolTip(title = NULL, desc = NULL, desc1 = desc, desc2 =
NULL, sub.special=TRUE)

Arguments

title 	 A title for the tool tip box. (optional)

desc 	 The text of the tool tip.

desc1 	 The first line of text of the tool tip.

desc2 	 The second line of text of the tool tip (used when the tool tip mode is 2.)

sub.special 	 If FALSE, do NOT replace special characters by their XML encodings (see details).

Details

This function sets tool tip text that will be used by the next graphics shape drawn with RSVGTipsDevice.

The RSVGTips device "consumes" the tooltip with the first shape that it draws, e.g., using rect, points, polygon, etc. If multiple objects are drawn (e.g., by supplying a vector to points), only the first shape drawn will have a tooltip. Note that lines does NOT use a tooltip.

Unless sub.special=FALSE, the following special characters in title, desc, etc. will be substituted by the XML entity as shown:

\& 	 &amp;

\' 	 &apos;

\" 	 &quot;

\< 	 &lt;

\> 	 &gt;

The exception to this is that an ampersand followed by lower-case letters and then a semi-colon is always left alone, because it already looks like an XML entity.

Value

Returns an invisible NULL.

Author(s)

Tony Plate tplate@acm.org

See Also

Examples here: RSVGTipsDevice

Examples


## Not run: setSVGShapeToolTip(title="A rectangle", desc="that is yellow")

[Package RSVGTipsDevice version 0.7-5 Index]