setSVGShapeURL(RSVGTipsDevice)	R Documentation

Set raw shape URL (XML text) for the next graphics shape drawn with the RSVGTips device.

Description

Set a URL to use as a hyperlink for the next graphics shape drawn with the RSVGTips device.

Usage


setSVGShapeURL(url, target=NULL, sub.special=TRUE)

Arguments

url 	 A full valid URL, e.g., "http://www.r-project.org"

target 	 The name of a target in the page - use "_blank" to make the link open a new browser tab or window

sub.special 	 If TRUE, substitue XML encodings for special characters in url and target. If FALSE, it is possible to produce an invalid XML file by supplying a URL with special characters (i.e., "<", ">", "&", and single- and double-quotes.)

Details

This function sets a hyperlink that will be included in the SVG commands output for the next graphics shape drawn with RSVGTipsDevice.

Value

Returns an invisible NULL.

Author(s)

Tony Plate tplate@acm.org

See Also

RSVGTipsDevice

Examples


## Not run: setSVGShapeURL("http://www.r-project.org")

[Package RSVGTipsDevice version 0.7-5 Index]